refactor(proactive): 迁出主动判断循环
