+++ agent/core/proactive_turn.py
@@ -19,7 +19,6 @@ ProactiveTurnPipeline — 主动回复链路顶层抽象。
 
 from __future__ import annotations
 
-import json
 import logging
 import random as _random_module
 import time
@@ -32,7 +31,7 @@ from agent.plugins.proactive_effects import (
     ProactiveEffectContext,
     ProactiveEffectProvider,
 )
-from agent.tool_hooks import ToolExecutionRequest, ToolExecutor, ToolHook
+from agent.tool_hooks import ToolExecutor, ToolHook
 from agent.turns.orchestrator import TurnOrchestrator
 from agent.turns.result import TurnResult
 from core.common.diagnostic_log import diagnostic_context, diagnostic_line
@@ -41,6 +40,7 @@ from proactive_v2.context import AgentTickContext
 from agent.core.drift_turn import DriftTurnPipeline
 from proactive_v2.gateway import DataGateway, GatewayDeps, GatewayResult
 from proactive_v2.modules_gate import GateResult, ProactiveGateChain
+from proactive_v2.modules_judge import ProactiveJudge
 from proactive_v2.modules_prompt import ProactivePromptBuilder
 from proactive_v2.modules_resolve import (
     ProactiveResolver,
@@ -50,7 +50,7 @@ from proactive_v2.modules_resolve import (
     ack_post_guard_fail,
     build_delivery_key,
 )
-from proactive_v2.tools import TOOL_SCHEMAS, ToolDeps, dispatch
+from proactive_v2.tools import ToolDeps
 
 logger = logging.getLogger(__name__)
 
@@ -177,6 +177,14 @@ class ProactiveTurnPipeline:
             recent_proactive_fn=self._recent_proactive_fn,
             ack_fn=self._tool_deps.ack_fn,
             alert_ack_fn=self._tool_deps.alert_ack_fn,
+        )
+        self._judge = ProactiveJudge(
+            cfg=self._cfg,
+            session_key=self._session_key,
+            llm_fn=self._llm_fn,
+            tool_deps=self._tool_deps,
+            tool_executor=self._tool_executor,
+            record_step_fn=self._record_tick_step,
         )
 
         # 1. drift_pipeline 的 step_recorder 指向本 pipeline 的记录方法。
@@ -489,70 +497,7 @@ class ProactiveTurnPipeline:
     # ── 3. Judge ──────────────────────────────────────────────────────
 
     async def _judge_evaluate(self, ctx: AgentTickContext, messages: list[dict]) -> None:
-        """LLM 多轮工具调用：逐条内容分类 → 草稿 → 收尾。"""
-
-        if self._llm_fn is None:
-            return
-
-        # 3.1 主 loop：模型自行决定调用工具，直到 finish_turn 或达到步数上限。
-        while ctx.steps_taken < self._cfg.agent_tick_max_steps:
-            ok = await self._run_tool_step(messages, ctx, loop_tag="loop", tool_choice="auto")
-            if not ok:
-                break
-            if ctx.terminal_action is not None:
-                break
-
-        # 3.2 完整性检查：如果 finish_skip 了但还有未分类条目，补全。
-        gw_result = self._last_gateway_result
-        if ctx.terminal_action == "skip" and gw_result is not None and gw_result.content_meta:
-            all_content_ids = {m["id"] for m in gw_result.content_meta}
-            classified_ids = ctx.interesting_item_ids | ctx.discarded_item_ids
-            unclassified_ids = all_content_ids - classified_ids
-            if unclassified_ids:
-                ctx.terminal_action = None
-                ctx.skip_reason = ""
-                ctx.skip_note = ""
-                titles_hint = "; ".join(
-                    f"{m['id']}（{m['title'][:40]}）"
-                    for m in gw_result.content_meta
-                    if m["id"] in unclassified_ids
-                )
-                completeness_msg = (
-                    f"【系统提示】以下 {len(unclassified_ids)} 个条目尚未完成分类：\n"
-                    f"{titles_hint}\n"
-                    "请对每条调用 mark_interesting 或 mark_not_interesting，"
-                    "全部分类完毕后再调用 message_push + finish_turn(decision=reply)，或 finish_turn(decision=skip, reason=...)。"
-                )
-                logger.info(
-                    "[proactive_v2] judge completeness: %d unclassified, resetting → %s",
-                    len(unclassified_ids),
-                    sorted(unclassified_ids),
-                )
-                messages.append({"role": "user", "content": completeness_msg})
-                for _ in range(5):
-                    if ctx.terminal_action is not None or ctx.steps_taken >= self._cfg.agent_tick_max_steps:
-                        break
-                    ok = await self._run_tool_step(messages, ctx, loop_tag="complete")
-                    if not ok:
-                        break
-
-        # 3.3 反思阶段：如果 interesting 已标好但还没 finish_turn，逼它收尾。
-        if ctx.terminal_action is None and ctx.interesting_item_ids and ctx.steps_taken < self._cfg.agent_tick_max_steps:
-            ids_str = ", ".join(sorted(ctx.interesting_item_ids))
-            reflection = (
-                f"【系统提示】你已将以下条目标记为 interesting：{ids_str}。\n"
-                "所有条目均已分类完毕。你必须现在调用 message_push 撰写推送，然后调用 finish_turn(decision=reply)；"
-                "或直接调用 finish_turn(decision=skip, reason=...)。不允许直接结束。"
-            )
-            logger.info("[proactive_v2] judge reflection: interesting=%d, injecting prompt", len(ctx.interesting_item_ids))
-            messages.append({"role": "user", "content": reflection})
-            for _ in range(3):
-                if ctx.terminal_action is not None or ctx.steps_taken >= self._cfg.agent_tick_max_steps:
-                    break
-                ok = await self._run_tool_step(messages, ctx, loop_tag="reflect", tool_choice="auto")
-                if not ok:
-                    break
-
+        await self._judge.evaluate(ctx, messages, self._last_gateway_result)
         self.last_ctx = ctx
 
     # ── 4. Resolve ────────────────────────────────────────────────────
@@ -590,156 +535,6 @@ class ProactiveTurnPipeline:
         )
         self._record_tick_log_finish(ctx)
         ctx.content_store.clear()
-
-    # ── LLM 工具单步 ──────────────────────────────────────────────────
-
-    async def _run_tool_step(
-        self,
-        messages: list[dict],
-        ctx: AgentTickContext,
-        *,
-        loop_tag: str,
-        tool_choice: str | dict = "auto",
-        schemas: list[dict] | None = None,
-    ) -> bool:
-        """用当前 messages 调一次模型，拿到本轮工具调用并执行。"""
-        active_schemas = schemas or TOOL_SCHEMAS
-        llm_fn = self._llm_fn
-        if llm_fn is None:
-            return False
-        tool_call = await llm_fn(messages, active_schemas, tool_choice)
-        if tool_call is None:
-            logger.warning(
-                "[proactive_v2] %s: llm_fn returned None at step %d, stopping",
-                loop_tag,
-                ctx.steps_taken,
-            )
-            return False
-        tool_name = tool_call.get("name", "")
-        tool_args = tool_call.get("input", {})
-        arg_summary = json.dumps(tool_args, ensure_ascii=False)[:200]
-        logger.info(
-            diagnostic_line(
-                "ProactiveTurnPipeline._run_tool_step",
-                event="tool_call",
-                flow="proactive",
-                phase="agent_loop",
-                session=self._session_key,
-                tick=ctx.tick_id,
-                action=str(tool_name or "-"),
-                counts=f"step:{ctx.steps_taken}",
-            )
-        )
-        logger.info(
-            "[proactive_v2] %s step %d: %s  args=%s",
-            loop_tag,
-            ctx.steps_taken,
-            tool_name,
-            arg_summary,
-        )
-        ctx.steps_taken += 1
-        exec_result = await self._tool_executor.execute(
-            ToolExecutionRequest(
-                call_id=str(tool_call.get("id") or f"call_{ctx.steps_taken}"),
-                tool_name=tool_name,
-                arguments=tool_args,
-                source="proactive",
-                session_key=self._session_key,
-            ),
-            lambda name, args: dispatch(name, args, ctx, self._tool_deps),
-        )
-        if exec_result.status == "error":
-            logger.warning(
-                diagnostic_line(
-                    "ProactiveTurnPipeline._run_tool_step",
-                    event="tool_result",
-                    flow="proactive",
-                    phase="agent_loop",
-                    session=self._session_key,
-                    tick=ctx.tick_id,
-                    action=str(tool_name or "-"),
-                    reason="tool_error",
-                    counts=f"step:{ctx.steps_taken}",
-                    note=str(exec_result.output)[:160],
-                )
-            )
-            logger.warning("[proactive_v2] %s: tool error: %s", loop_tag, exec_result.output)
-            result = str(exec_result.output)
-            call_id = tool_call.get("id") or f"call_{ctx.steps_taken}"
-            self._record_tick_step(
-                ctx,
-                phase=f"{loop_tag}:error",
-                tool_name=tool_name,
-                tool_call_id=str(call_id),
-                tool_args=tool_args,
-                tool_result_text=result,
-            )
-            self._append_tool_messages(
-                messages,
-                tool_name=tool_name,
-                tool_args=tool_args,
-                tool_call_id=call_id,
-                result=result,
-            )
-            return False
-        result = str(exec_result.output)
-        logger.info(
-            diagnostic_line(
-                "ProactiveTurnPipeline._run_tool_step",
-                event="tool_result",
-                flow="proactive",
-                phase="agent_loop",
-                session=self._session_key,
-                tick=ctx.tick_id,
-                action=str(tool_name or "-"),
-                reason="-",
-                counts=f"step:{ctx.steps_taken}",
-            )
-        )
-        call_id = tool_call.get("id") or f"call_{ctx.steps_taken}"
-        self._record_tick_step(
-            ctx,
-            phase=loop_tag,
-            tool_name=tool_name,
-            tool_call_id=str(call_id),
-            tool_args=tool_args,
-            tool_result_text=result,
-        )
-        self._append_tool_messages(
-            messages,
-            tool_name=tool_name,
-            tool_args=tool_args,
-            tool_call_id=call_id,
-            result=result,
-        )
-        return True
-
-    @staticmethod
-    def _append_tool_messages(
-        messages: list[dict],
-        *,
-        tool_name: str,
-        tool_args: dict,
-        tool_call_id: str,
-        result: str,
-    ) -> None:
-        messages.append({
-            "role": "assistant",
-            "content": f"调用工具 {tool_name}",
-            "tool_calls": [{
-                "id": tool_call_id,
-                "type": "function",
-                "function": {
-                    "name": tool_name,
-                    "arguments": json.dumps(tool_args, ensure_ascii=False),
-                },
-            }],
-        })
-        messages.append({
-            "role": "tool",
-            "tool_call_id": tool_call_id,
-            "content": result,
-        })
 
     # ── Tick 日志记录 ──────────────────────────────────────────────────
